app/requirements.txt
fastapi
uvicorn
jinja2
tensorflow
pillow